Implement symptom entry functionality with UI and ViewModel integration

+++ shared/src/commonMain/kotlin/com/example/symptomtracker/presentation/SymptomEntryState.kt
@@ -2,6 +2,6 @@ package com.example.symptomtracker.presentation
 
 data class SymptomEntryState(
   val name: String = "",
-  val severity: Int = 5,
+  val severity: Float = 5f,
   val notes: String = "",
 )
